{"types":"C","runs":6}

+++ Problem24.java
@@ -3,9 +3,9 @@ import acm.program.ConsoleProgram;
 //24. მომხმარებელს შეყავს მთელი რიცხვი n, პროგრამამ უნდა დაბეჭდოს n ის გამყოფების რაოდენობა 
 public class Problem24 extends ConsoleProgram {
 	public void run(){
-		int n = readInt("Give me a number: ");
+		double n = readInt("Give me a number: ");
 		int division = 0;
-		for(int i = 0; i < n; i++){
+		for(double i = 0; i < n; i++){
 			if(n % i == 0){
 			division += 1;
 			}
